\name{NS}
\alias{NS}
\docType{data}
\title{RNAseq data from paired flow-sorted normal/treated PBSC samples}
\description{
This data package contains a \code{KallistoExperiment} instance as well
as 6 HDF5 files with 100 bootstraps each from an RNAseq experiment in HSPCs.
}
\usage{data(NS)}
\examples{
data(NS)  
show(NS)
}
\keyword{datasets}
